Rename OXT to O for dssp
DSSP can sometime trip on the C terminal residue of a protein, ignoring
it because the oxygen is named OXT. The error is not systematic, but
when it happens it causes the last residue not to appear on the output
PDB file. The residue still appear in the ITP file causing a mismatch.

This commit renames the OXT atom to O in the structure that is passed to
DSSP.