Refactor how force fields are loaded
This commit replaces the way force fields are loaded by a meta class.
Before this commit, a tricky gymnastic was done to find the python file
named '<force-field>_ff.py' in various places. It implied some fragile
introspection and required to bypass the way modules are loaded.

This commit replaces this by a ForceFields meta class that register all
its children in the `forcefield.FORCE_FIELD_COLLECTION` dictionary.

Hopefully, this will allow to count the force fields in the coverage
report.