Split the force field init
Move most of the init logic of the force field in a method called setup.
Remove the call to finish from that method. Add an __init__ to the base
class that calls setup and finish.

The point is to latter add a set_mapping method that will only have to
be overloaded in rare cases where the mapping change.