Allows to describe a force field as a JSON file
A JSON file can now be passed to the `-ff` option. The JSON file must
describe the force field in the same way as the setup method of a force
field. No sanity check is performed so it is let to the user to make
sure that the JSON file defines all the required fields.

The 'special' field uses a tuple as key, which is not permitted in JSON
so the format had to be changed. In a setup method, the 'special'
attribute has to be formatted as
    {(("SC1","CYS"), ("SC1","CYS")): [0.24, null]}
in the JSON file, the same attribute is formatted as
    [["SC1","CYS"], ["SC1","CYS"], [0.24, null]]